Small adjustments

+++ src/genetic/judgment/JN5.java
@@ -7,13 +7,13 @@ import bidders.BiddingAgent.Attitude;
 import genetic.Node;
 
 public class JN5 extends JudgmentNode {
-	private final float timeStepPercentageThreshold = (new Random().nextInt(91) + 5) / 100.f; // Percentage of passed time steps after which to take destination2
+	private final double timeStepPercentageThreshold = (new Random().nextInt(91) + 5) / 100.; // Percentage of passed time steps after which to take destination2
 	
 	@Override
 	public Node makeDecision(Auction auction, Attitude attitude) {
 		checkDestinationNodes();
 		
-		float timeStepPercentage = (float) (auction.getCurrentTimeStep() / auction.getTotalTimeSteps() * 10);
+		double timeStepPercentage = auction.getCurrentTimeStep() / (auction.getTotalTimeSteps() * 10.);
 		
 		return (timeStepPercentage <= timeStepPercentageThreshold) ? destination1 : destination2;
 	}
